Create client with all requests to GCP and refactor GCS
