+++ pkg/google/client/billing_test_helpers.go
@@ -1,4 +1,4 @@
-package billing
+package client
 
 import (
 	"context"
