+++ pkg/google/client/bucket_test.go
@@ -1,4 +1,4 @@
-package gcs
+package client
 
 import (
 	"context"
@@ -7,6 +7,7 @@ import (
 	"testing"
 
 	"cloud.google.com/go/storage"
+	"github.com/grafana/cloudcost-exporter/pkg/google/client/cache"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/api/option"
@@ -24,7 +25,7 @@ func TestNewBucketClient(t *testing.T) {
 	}
 	for name, test := range tests {
 		t.Run(name, func(t *testing.T) {
-			client := NewBucketClient(test.client)
+			client := newBucket(test.client, cache.NewNoopCache[[]*storage.BucketAttrs]())
 			if client == nil {
 				t.Errorf("expected cloudCatalogClient to be non-nil")
 			}
@@ -77,7 +78,7 @@ func TestBucketClient_List(t *testing.T) {
 			for _, project := range test.projects {
 				sc, err := storage.NewClient(context.Background(), option.WithEndpoint(test.server.URL), option.WithAPIKey("hunter2"))
 				require.NoError(t, err)
-				bc := NewBucketClient(sc)
+				bc := newBucket(sc, cache.NewNoopCache[[]*storage.BucketAttrs]())
 				got, err := bc.List(context.Background(), project)
 				assert.Equal(t, test.wantErr, err != nil)
 				assert.NotNil(t, got)
