+++ pkg/google/client/cache/bucket_cache_test.go
@@ -1,4 +1,4 @@
-package gcs
+package cache
 
 import (
 	"strconv"
